Refactor the page container component
- Introduce a PageContentContainer that provides a single source of
  truth for content containers (we have three: header, body, and
  footer).

- Move size-related utilities into the design tokens directory and
  rename it as "spacing".

- Add padding to the content containers.
