+++ src/design/spacing.js
@@ -1,0 +1,9 @@
+/*
+ * Spaced-related tokens.
+ */
+
+/* Max width of the content container */
+export const mainWidth = '75rem';
+
+export const regularMarginRem = 1;
+export const mediumMarginRem = 2;
