--- src/utilities/sizes.js
@@ -1,8 +1,0 @@
-/*
- * Sizing-related utilities for Styled Components.
- */
-
-export const mainWidthRem = 75;
-
-export const regularMarginRem = 1;
-export const mediumMarginRem = 2;
